Profile Page Dark Mode bug fix
SHADCN UI was setting theme to dark and all the places where we used material ui was not updating the color , so i fixed the issue by hardcoding the material ui components text fields to black and the background to white always.

This was the quickest and most convenient fix that i could find given the time constraint of developing the application.